+++ lib/ecspanse.ex
@@ -31,7 +31,7 @@ defmodule Ecspanse do
 
   ```elixir
   defmodule TestServer1 do
-    use Ecspanse, fps_limit: 60
+    use Ecspanse, fps_limit: 60, version: 1
 
     def setup(data) do
       data
@@ -51,19 +51,22 @@ defmodule Ecspanse do
   ## Configuration
 
   The following configuration options are available:
-  - `:fps_limit` (optional) - the maximum number of frames per second. Defaults to :unlimited.
+  - `:fps_limit` (optional) - integer or :unlimited - the maximum number of frames per second. Defaults to :unlimited.
+  - `:version` (optional) - non negative integer - the version of the ECS game logic.
+  Can be used to determine backwards compatibility when saving and loading state(see `Ecspanse.Snapshot` for details).
+  Defaults to 0.
 
 
   # Special Resources
 
-  Some special resources, such as `State` or `FPS`, are created by default by the framework.
+  Some special resources, such as `FPS`, are created by default by the framework.
 
   """
 
-  require Logger
-
   alias __MODULE__
   alias Ecspanse.Util
+
+  require Logger
 
   defmodule Data do
     @moduledoc """
@@ -106,10 +109,12 @@ defmodule Ecspanse do
   end
   ```
   """
-  @callback setup(Ecspanse.Data.Ecspanse.Data.t()) :: Ecspanse.Data.Ecspanse.Data.t()
+  @callback setup(Ecspanse.Data.t()) :: Ecspanse.Data.t()
 
   defmacro __using__(opts) do
     quote bind_quoted: [opts: opts], location: :keep do
+      @behaviour Ecspanse
+
       import Ecspanse,
         only: [
           insert_resource: 2,
@@ -126,17 +131,23 @@ defmodule Ecspanse do
           add_system_set: 3
         ]
 
-      @behaviour Ecspanse
-
       fps_limit = Keyword.get(opts, :fps_limit, :unlimited)
+      version = Keyword.get(opts, :version, 0)
 
       if fps_limit && not (is_integer(fps_limit) || fps_limit == :unlimited) do
         raise ArgumentError,
-              "If set, the option :fps_limit must be a non negative integer in the Server module #{Kernel.inspect(__MODULE__)}"
+              "If set, the option :fps_limit must be a non negative integer in the module #{Kernel.inspect(__MODULE__)}"
+      end
+
+      unless is_integer(version) do
+        raise ArgumentError,
+              "The option :version must be an integer in the module #{Kernel.inspect(__MODULE__)}"
       end
 
       Module.register_attribute(__MODULE__, :fps_limit, accumulate: false)
       Module.put_attribute(__MODULE__, :fps_limit, fps_limit)
+      Module.register_attribute(__MODULE__, :ecs_version, accumulate: false)
+      Module.put_attribute(__MODULE__, :ecs_version, version)
 
       # THIS WILL BE THE MIX ENV OF THE PROJECT USING ECSPANSE, NOT ECSPANSE ITSELF
       if Mix.env() == :test do
@@ -146,7 +157,8 @@ defmodule Ecspanse do
           if arg == :test do
             payload = %{
               ecspanse_module: __MODULE__,
-              fps_limit: @fps_limit
+              fps_limit: @fps_limit,
+              ecs_version: @ecs_version
             }
 
             %{
@@ -167,7 +179,8 @@ defmodule Ecspanse do
         def child_spec(_arg) do
           payload = %{
             ecspanse_module: __MODULE__,
-            fps_limit: @fps_limit
+            fps_limit: @fps_limit,
+            ecs_version: @ecs_version
           }
 
           %{
@@ -241,11 +254,7 @@ defmodule Ecspanse do
 
   @spec add_frame_start_system(Ecspanse.Data.t(), system_module :: module(), opts :: keyword()) ::
           Ecspanse.Data.t()
-  def add_frame_start_system(
-        %Ecspanse.Data{operations: operations} = data,
-        system_module,
-        opts \\ []
-      ) do
+  def add_frame_start_system(%Ecspanse.Data{operations: operations} = data, system_module, opts \\ []) do
     opts = merge_system_options(opts, data.system_set_options)
 
     if Keyword.get(opts, :run_after) do
@@ -255,8 +264,7 @@ defmodule Ecspanse do
     end
 
     system =
-      %Ecspanse.System{module: system_module, queue: :frame_start_systems, execution: :sync}
-      |> add_run_conditions(opts)
+      add_run_conditions(%Ecspanse.System{module: system_module, queue: :frame_start_systems, execution: :sync}, opts)
 
     %Ecspanse.Data{data | operations: [{:add_system, system} | operations]}
   end
@@ -274,7 +282,7 @@ defmodule Ecspanse do
   ## Order of execution
 
   Systems are executed each frame during the game loop. Sync systems run in the order they were added to the data's operations list.
-  Async systems are grouped in batches depending on the componets they are locking.
+  Async systems are grouped in batches depending on the components they are locking.
   See the `Ecspanse.System` module for more information about component locking.
 
   The order in which async systems run can pe specified using the `run_after` option.
@@ -319,7 +327,7 @@ defmodule Ecspanse do
     end
     ```
 
-  It is important to nothe that the run conditions are evaluated only once per frame, at the beginning of the frame.
+  It is important to note that the run conditions are evaluated only once per frame, at the beginning of the frame.
   So any change in the running conditions will be picked up in the next frame.
 
   If the system is part of a system set, and both the system and the system set have run conditions, the conditions are cumulative.
@@ -351,13 +359,10 @@ defmodule Ecspanse do
       end
 
     system =
-      %Ecspanse.System{
-        module: system_module,
-        queue: :batch_systems,
-        execution: :async,
-        run_after: run_after
-      }
-      |> add_run_conditions(opts)
+      add_run_conditions(
+        %Ecspanse.System{module: system_module, queue: :batch_systems, execution: :async, run_after: run_after},
+        opts
+      )
 
     %Ecspanse.Data{data | operations: [{:add_system, system} | operations]}
   end
@@ -380,11 +385,7 @@ defmodule Ecspanse do
   """
   @spec add_frame_end_system(Ecspanse.Data.t(), system_module :: module(), opts :: keyword()) ::
           Ecspanse.Data.t()
-  def add_frame_end_system(
-        %Ecspanse.Data{operations: operations} = data,
-        system_module,
-        opts \\ []
-      ) do
+  def add_frame_end_system(%Ecspanse.Data{operations: operations} = data, system_module, opts \\ []) do
     opts = merge_system_options(opts, data.system_set_options)
 
     if Keyword.get(opts, :run_after) do
@@ -394,8 +395,7 @@ defmodule Ecspanse do
     end
 
     system =
-      %Ecspanse.System{module: system_module, queue: :frame_end_systems, execution: :sync}
-      |> add_run_conditions(opts)
+      add_run_conditions(%Ecspanse.System{module: system_module, queue: :frame_end_systems, execution: :sync}, opts)
 
     %Ecspanse.Data{data | operations: [{:add_system, system} | operations]}
   end
@@ -559,8 +559,7 @@ defmodule Ecspanse do
       end
 
     event_payload =
-      event_payload
-      |> Keyword.put(:inserted_at, System.os_time())
+      Keyword.put(event_payload, :inserted_at, System.os_time())
 
     {{event_module, key}, struct!(event_module, event_payload)}
   end
@@ -576,9 +575,8 @@ defmodule Ecspanse do
 
   # merge the system options with the system set options
   # this is the reason the conditional run states are a list
-  defp merge_system_options(system_opts, system_set_opts)
-       when is_list(system_opts) and is_map(system_set_opts) do
-    system_set_opts = Map.values(system_set_opts) |> List.flatten() |> Enum.uniq()
+  defp merge_system_options(system_opts, system_set_opts) when is_list(system_opts) and is_map(system_set_opts) do
+    system_set_opts = system_set_opts |> Map.values() |> List.flatten() |> Enum.uniq()
 
     (system_opts ++ system_set_opts)
     |> Enum.group_by(fn {k, _v} -> k end, fn {_k, v} -> v end)
@@ -620,7 +618,8 @@ defmodule Ecspanse do
   defp conditional_run_state(opts, option) do
     # This is a list. See the `merge_system_options/2` function for more info.
     # The run state conditions from systems and system sets are cumulative.
-    Keyword.get(opts, option, [])
+    opts
+    |> Keyword.get(option, [])
     |> Enum.map(fn
       {state_module, state} when is_atom(state_module) and is_atom(state) ->
         unless is_atom(state) and state in state_module.__states__() do