+++ lib/ecspanse/template/component/timer.ex
@@ -56,13 +56,11 @@ defmodule Ecspanse.Template.Component.Timer do
   end
   ```
 
-  See [a working example](./tutorial.md#energy-regeneration) in the tutorial
+  [A working example](./tutorial.md#energy-regeneration) is available in the tutorial
 
   """
-  @timer_component_tag :ecs_timer
-
   use Ecspanse.Template.Component,
-    tags: [@timer_component_tag],
+    tags: [:ecs_timer],
     state: [:duration, :time, :event, mode: :repeat, paused: false]
 
   @mode [:repeat, :once, :temporary]
@@ -72,11 +70,8 @@ defmodule Ecspanse.Template.Component.Timer do
     with :ok <- validate_duration(state[:duration]),
          :ok <- validate_time(state[:time]),
          :ok <- validate_event(state[:event]),
-         :ok <- validate_mode(state[:mode]),
-         :ok <- validate_paused(state[:paused]) do
-      :ok
-    else
-      {:error, reason} -> {:error, reason}
+         :ok <- validate_mode(state[:mode]) do
+      validate_paused(state[:paused])
     end
   end
 
@@ -126,5 +121,5 @@ defmodule Ecspanse.Template.Component.Timer do
   end
 
   @doc false
-  def timer_component_tag, do: @timer_component_tag
+  def timer_component_tag, do: :ecs_timer
 end